requirements.txt
flask==1.1.4
flask-cors==3.0.10
flask-bcrypt==0.7.1
sklearn
pymongo==3.12.1
pandas==1.3.3
numpy==1.21.2
scipy==1.7.1
requests==2.2.0
pyjwt==2.3.0
jwt==1.3.1
pyodbc==4.0.32
azure-storage-blob
urllib3
dnspython